backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
twilio==9.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
eventlet==0.35.1
pymongo==4.6.1
dnspython==2.4.2
